refactor - refurbishing subRedditContext for Form and Heatmap compos
