+++ src/components/subRedditForm/subredditForm.style.js
@@ -1,4 +1,4 @@
-import styled, { keyframes } from 'styled-components/macro';
+import styled from 'styled-components/macro';
 
 export const Container = styled.section`
   width: 100vw;
@@ -49,19 +49,4 @@ export const Input = styled.input`
 export const PrefixedText = styled.p`
   font-size: ${({ theme }) => theme.font.size.h6};
   color: ${({ theme }) => theme.color.lighterGray};
-`;
-
-const rotate360 = keyframes`
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-`;
-
-export const Spinner = styled.img`
-  animation: ${rotate360} 1.5s linear infinite;
-  width: 70px;
-  margin: 1rem 0;
 `;
