+++ src/pages/searchpage.js
@@ -1,11 +1,12 @@
 import React from 'react';
-import { SubRedditForm } from '../components/index';
+import { SubRedditForm, Heatmap } from '../components/index';
 import SubRedditContextProvider from '../contexts/subRedditContext';
 
 export default function SearchPage() {
   return (
     <SubRedditContextProvider>
       <SubRedditForm />
+      <Heatmap />
     </SubRedditContextProvider>
   );
 }
